Fire and Skybox
I added several more fires. I didnt want to add to many until we had the
flightpath determed but it is a good spread arount the town. I also used
a different skybox thats in a dome so that that there isnt just a void
after the town. When editing turning off the citydoom and skydoom makes
it easier to see the fire.