Fire/Water Interactions Enabled
DeathByWater script allows fire assets to have a health system with
damage by water. Water layer allows for water detection by fire.
Colliders added to fire for water particle collision detection.

Fire assets resistance can be configured via the DeathByWater script.

Some fire asset prefabs had to be reverted and modified in the scene in
order to restore proper prefab binding and allow the colliders to
properly represent the bounds of the fire.

Water spray toggle enabled with primary (Left) mouse button held-down.

Water spray particle system configured.

These interactions are in a rudimentary form and are subject to change
and tuning.

FireLayer simply fixes a bug in the scripts that came with the fire
assets.

Custom Fluvio Particle - FireFrenzy is a custom water material shader
for this scene.